+++ database/migrations/create/2020_06_20_000000_create_orders_payment_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             Schema::createIfMissing( 'nexopos_orders_payments', function ( Blueprint $table ) {
                 $table->bigIncrements( 'id' );
                 $table->integer( 'order_id' );
-                $table->float( 'value', 18, 5 )->default( 0 );
+                $table->decimal( 'value', 18, 5 )->default( 0 );
                 $table->integer( 'author_id' );
                 $table->string( 'identifier' );
                 $table->string( 'uuid' )->nullable();
