+++ database/migrations/create/2020_06_20_000000_create_registers_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
                 $table->text( 'description' )->nullable();
                 $table->integer( 'used_by' )->nullable();
                 $table->integer( 'author_id' );
-                $table->float( 'balance', 18, 5 )->default( 0 );
+                $table->decimal( 'balance', 18, 5 )->default( 0 );
                 $table->string( 'uuid' )->nullable();
                 $table->timestamps();
             } );
