+++ database/migrations/create/2021_01_23_225713_create_customers_rewards_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->integer( 'customer_id' );
             $table->integer( 'reward_id' );
             $table->string( 'reward_name' );
-            $table->float( 'points', 18, 5 );
-            $table->float( 'target', 18, 5 );
+            $table->decimal( 'points', 18, 5 );
+            $table->decimal( 'target', 18, 5 );
             $table->timestamps();
         } );
     }
